account-service/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
minio==7.2.7
kafka-python==2.0.2
drf-spectacular==0.27.2
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
black==24.4.2
isort==5.13.2
flake8==7.1.0
mypy==1.10.0
django-stubs==5.0.2
pre-commit==3.7.1
coverage==7.5.3
